requirements.txt
dvclive>=3.0
pandas
pyaml
scikit-learn>=1.3
scipy
matplotlib
torch==1.12.1
transformers==4.18.0
numpy==1.21.5
tqdm==4.64.0
ipywidgets==8.0.5